Added an album list view.
This is far from ideal since compilation albums are listed multiple
times (once per artist). I'm not sure yet how we want to deal with this.
It is straightforward to include artist name alongside album name for
cases where there really are two albums with the same name, say
"Greatest Hits", but we do need a way to identify compilation albums
somehow and group the tracks accordingly.

Perhaps this patch is not the best solution.